So about this Dirty Laundry. It would appear that with IOS7, we are now seeing occasions where sometimes an exception is thrown, deep in autorelease pool. The exception complains about   _overrideInputViewNextResponderWithResponder:]: unrecognized selector sent to instance
Whilst trying to dealloc, and [UITableViewCell removeFromSuperview].

The dirty laundry hack in IBA, seems to add all the cells onto the hidden cache as the controller's view is removed (i.e. close the form), which I believe was not the intent.

Adding this additional call to clearHiddenCellCache after the controller's view has disappeared, should be safe.